% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{addreturns}
\alias{addreturns}
\title{Adds various returns as well as additional statistics.}
\usage{
addreturns(df)
}
\arguments{
\item{df}{Dataframe pulled from alphavantager}
}
\value{
A more detailed dataframe with additiona return metrics and summary statistics.
}
\description{
Adds percentage returns, interday returns, total returns, cumulative returns, multiplicative returns, range, and dollar returns to the default dataframe pulled from alphavantager
}
\examples{
addreturns(daily_data_from_av)
}
