% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{candles}
\alias{candles}
\title{Candlestick chart}
\usage{
candles(df)
}
\arguments{
\item{df}{Dataframe with price data.}
}
\value{
A candlestick  chart
}
\description{
Returns plot_ly candlestick chart
}
\examples{
candles(SPYdaily)
candles(SPY60)
candles(SPYwk)
}
