% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{compare_returns}
\alias{compare_returns}
\title{Compare returns visually between two securities}
\usage{
compare_returns(a, b, a_name, b_name)
}
\arguments{
\item{a}{The first dataframe filled with data from the function getdaily()}

\item{b}{The second dataframe filled with data from the function getdaily()}

\item{a_name}{Character string name for first dataframe}

\item{b_name}{Character string name for second dataframe}
}
\value{
Returns plotly graph comparing returns, cumulative nominal returns, and cumulative multiplicative returns. Click graph name to un-toggle for better visibility.
}
\description{
Prints a plotly graph comparing returns, cumulative nominal returns, and cumulative multiplicative returns
}
\examples{
compare_returns(GMEdaily, SPYdaily, "GME", "SPY")
}
