% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{crypto60}
\alias{crypto60}
\title{Get Cryptocurrency Data at the hourly level localized to current time zone}
\usage{
crypto60(coin_name)
}
\arguments{
\item{coin_name}{The ticker symbol for the concurrency as a string}
}
\value{
A data frame with daily data such as the high, low, open, close, and associated returns. Available in the global environment.
Adjusted to local time zone.
}
\description{
Summary statistics for the movements on the hourly level
}
\examples{
crypto60("BTC")
}
