% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{cryptodaily}
\alias{cryptodaily}
\title{Get Cryptocurrency Data at the Daily Level}
\usage{
cryptodaily(coin_name)
}
\arguments{
\item{coin_name}{The ticker symbol for the concurrency as a string}
}
\value{
A data frame with daily data such as the high, low, open, close, and associated returns. Available in the global environment.
}
\description{
Daily, as in the summary statistics for the daily movement
}
\examples{
cryptodaily("WTI")
}
