% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{fastzoom}
\alias{fastzoom}
\title{Fast Zoom}
\usage{
fastzoom(plot, x)
}
\value{
The graph but zoomed in where you want to, mostly for daily data
}
\description{
An Easy way to get to where you want to on a candlestick chart or other plots
}
\examples{
fastzoom(candles(SPYdaily), "20220202")
}
