% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{fibs}
\alias{fibs}
\title{Show Fibonacci bands}
\usage{
fibs(df, showgraph = TRUE, title = NULL)
}
\arguments{
\item{df}{Dataframe with price data, works with various intervals}

\item{showgraph}{Whether or not you want the function to pop out the visual}

\item{title}{A character string for the Title of your graph}
}
\value{
Returns graph with various levels as well as a vector
}
\description{
Displays Fibonacci bands
}
\examples{
fibs(tail(SPYdaily, 200))
SPYdailyfibs <- fibs(tail(SPYdaily, 200))
fibs(SPY15)
}
