% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{findcor}
\alias{findcor}
\title{Calculate the correlation between a column shared within two dataframes.}
\usage{
findcor(a, b, sdata)
}
\arguments{
\item{a}{First dataframe}

\item{b}{Second dataframe}

\item{sdata}{A string of the column name to compare}
}
\value{
The correlation as a single numeric.
}
\description{
Two dataframes from one of the "get" functions recommended, but works with any dataframe that shares columns as well as column lengths.
}
\examples{
findcor(SPYdaily, GMEdaily, "returns")
}
