% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{genMA}
\alias{genMA}
\title{Generate moving averages}
\usage{
genMA(df, ma)
}
\arguments{
\item{df}{Dataframe with price data}

\item{ma}{# of periods for the moving average to calculate}
}
\value{
a vector with the same number of columns as df showing the moving averages. Periods before moving
average should be not considered for use. Output is kept same columns for compatability.
}
\description{
Description
}
\examples{
SPYDMA200 <- genMA(SPYdaily, 14)
}
