% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{getdaily}
\alias{getdaily}
\title{Get Stock Data at the Daily Level}
\usage{
getdaily(ticker)
}
\arguments{
\item{ticker}{The ticker symbol as a string}
}
\value{
A data frame with daily data such as the high, low, open, close, and associated returns. Available in the global environment.
}
\description{
Daily, as in the summary statistics for the daily movement
}
\examples{
\dontrun{
getdaily("SPY")
}
}
