% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{getweekly}
\alias{getweekly}
\title{Get Stock Data at the Weekly level}
\usage{
getweekly(ticker)
}
\arguments{
\item{ticker}{The ticker symbol as a string}
}
\value{
A data frame with daily data such as the high, low, open, close, and associated returns. Available in the global environment.
}
\description{
Summary statistics for the stock movements on the weekly
}
\examples{
getweekly("WTI")
}
