% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\docType{data}
\name{GMEdaily}
\alias{GMEdaily}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 360 rows and 13 columns.
}
\source{
Alpha Vantager API
}
\usage{
GMEdaily
}
\description{
This is data to be included in my package
}
\references{
\url{https://www.alphavantage.co/documentation/}
}
\keyword{datasets}
