% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{idret}
\alias{idret}
\title{After-hour and Pre-market Returns}
\usage{
idret(df)
}
\arguments{
\item{df}{Dataframe with daily data}
}
\value{
A vector in dataframe format of combined after-hour, overnight, and pre-market returns in percentage
}
\description{
Show returns between the close of the last trading day and the open of the current trading day
}
\examples{
idret(SPYdaily)
}
