% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{nvi}
\alias{nvi}
\title{Negative Volume Index}
\usage{
nvi(df)
}
\arguments{
\item{df}{Dataframe with price data.}
}
\value{
Returns a 1 x # of columns in df dataframe
}
\description{
Calculates the  negative volume index, uses closing price of the time period
}
\examples{
nvi(tail(SPYdaily, 200))
}
