% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{ret_to_cr}
\alias{ret_to_cr}
\title{Total return to cumulative return}
\usage{
ret_to_cr(list_of_returns)
}
\arguments{
\item{list_of_returns}{Vector in dataframe showing returns}
}
\value{
A vector in dataframe format cumulative percentage returns
}
\description{
The cumulative percentage is the addition of subsequent total daily returns.
}
\examples{
ret_to_cr(SPYdaily$returns)
}
