% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{showidreturns}
\alias{showidreturns}
\title{Frequency plot based on Absolute percentage movements}
\usage{
showidreturns(df, name_in_string)
}
\arguments{
\item{df}{Dataframe from the "get"}

\item{name_in_string}{Name of security associated with dataframe}
}
\value{
A plot_ly graph showing the frequency of absolute returns.
"All Data" - no data omitted
"Cumulative Probability" Cumulative frequency graph of returns
"No Extremes" filter out skewed data
}
\description{
Shows the cumulative probabilities of each percentage movement
}
\examples{
showidreturns(SPYdaily, "SPY")
}
