% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{streak}
\alias{streak}
\title{Streak}
\usage{
streak(df)
}
\arguments{
\item{df}{Dataframe with price data.}
}
\value{
Returns a 1 x # of columns in df dataframe
}
\description{
Counts the number of days the open price has moved consecutively
Negative and positive streaks are represented by their sign
}
\examples{
streak(tail(SPYdaily, 200))
}
