% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{streak_var}
\alias{streak_var}
\title{Streak (Multiple Variables)}
\usage{
streak_var(df, var)
}
\arguments{
\item{df}{Dataframe with price data.}

\item{var}{String of column name you wish to see streak in}
}
\value{
Returns a 1 x # of columns in df dataframe
}
\description{
Counts the number of days the open price has moved consecutively
Negative and positive streaks are represented by their sign; only works with various types of returns or mutated vector created by diff()
}
\examples{
streak2(tail(SPYdaily,200), "tot_ret")
}
