% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{trtomr}
\alias{trtomr}
\title{Total return to multiplicative return}
\usage{
trtomr(df)
}
\arguments{
\item{df}{Dataframe with daily data}
}
\value{
A vector in dataframe format cumulative multiplicative returns
}
\description{
Multiplicative returns are always comparative to the earliest return
}
\examples{
trtomr(SPYdaily)
}
