% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{volatility_freq}
\alias{volatility_freq}
\title{Frequency plot of Range, as well as maximum Upward and Downward Movement}
\usage{
volatility_freq(df, tick_name, cumulative = FALSE)
}
\arguments{
\item{df}{Dataframe with daily data}

\item{tick_name}{The ticker so the graph is correct}

\item{cumulative}{Default is FALSE, turn to TRUE for a cumulative plot.}
}
\value{
Frequency plot where you can find intraday volatility (range), maximun upside (Upward Movement), maximun downside (Downward Movement) on a cumulative percentile basis
}
\description{
a plot_ly plot
}
\examples{
volatilty_freq_cum (SPYdaily, "SPY")
}
