% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{volatilty_freq_cum}
\alias{volatilty_freq_cum}
\title{Cumulative Frequency plot of Range, as well as maximun Upward and Downward Movement}
\usage{
volatilty_freq_cum(df, tick_name)
}
\arguments{
\item{df}{Dataframe with daily data}

\item{tick_name}{The ticker so the graph is correct}
}
\value{
Cumulative frequency plot where you can find intraday volatility (range), maximun upside (Upward Movement), maximun downside (Downward Movement) on a cumulative percentile basis
}
\description{
a plot_ly plot
}
\examples{
volatilty_freq(SPYdaily, "SPY")
}
