% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{volp}
\alias{volp}
\title{Relative percentage from the Maximum Value}
\usage{
volp(df)
}
\arguments{
\item{df}{Dataframe with price data.}
}
\value{
Returns a 1 x # of columns in df dataframe in percentage
}
\description{
Returns the percent of the maximum volume movement from data
}
\examples{
volp(tail(SPYdaily, 200))
}
