% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{volume_analysis}
\alias{volume_analysis}
\title{Graph of Volume Indicators}
\usage{
volume_analysis(df, nameof)
}
\arguments{
\item{df}{Dataframe with price data.}
}
\value{
Returns plot_ly graph with PVI, NVI, Open and % Max Volatility
}
\description{
Follows Open, PVI,  NVI, and %Max Volatility
}
\examples{
volume_analysis(SPY15, "SPY")
}
