% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avfintools.R
\name{project}
\alias{project}
\title{Projects future prices based on regression}
\usage{
project(df, tickername)
}
\arguments{
\item{df}{Dataframe with price data. The opening price is used for projection purposes, works for all security types}

\item{tickername}{The ticker or the security you are putting in.}
}
\value{
What comes out of this function
}
\description{
Works with multiple time intervals. Do not oversupply with data, regression is expensive.
}
\examples{

project(tail(SPYdaily,200), "SPY")
project(SPY15, "SPY")
}
